+++ bin/setup_db.dart
@@ -509,15 +509,7 @@ Future<void> insertSccReportIsone() async {
 }
 
 Future<void> insertTccClearedPricesNyiso() async {
-  var config = ComponentConfig(
-      host: 'localhost:27017',
-      dbName: 'nyiso',
-      collectionName: 'tcc_clearing_prices');
-  var dir =
-      '${env['HOME']!}/Downloads/Archive/Nyiso/TCC/ClearingPrices/ToProcess/';
-  var archive = nyiso_tcc_cp.NyisoTccClearingPrices(config: config)..dir = dir;
-  // await archive.setupDb();
-
+  var archive = getNyisoTccClearingPriceArchive();
   await archive.db.open();
   var files = Directory(archive.dir).listSync().whereType<File>();
   // var files = Directory(archive.dir).listSync().whereType<File>();